Height of menu is an issue.
After debugging: We need to find a way to tell if from 5 possible headers in the list, the current shown first 5 rows contain how many of those 5 headers. Else the height will be too high as it always bases on the total 5 headers in teh list (e.g. list got 20 rows, and only 5 currentl show containing 1 or 2 headers), allthough only 1 or 2 headers are currently shown.