Added IsJustaGhost 20240420 changes (enabled/..)
Added improved OnGlobalMouseUp functionality.
Added item.enaled to processNameString and updateLabelsStrings, for if it is a function, it is updated the same as name and label.
Added dynamic selectable item based on control.selectable and has callback
Only closeOnSelect controls will allow highlighting on mouse over if has callback
Added inhanced UpdateHeight function.
Fixed name of widthPadding in row template
Move RefreshSortedItems to comboBox_base where entries are acquired by self:GetEntries() in submenuClass and contextMenuClass
Moved exposedVariables and exposedFunctions in their locals and, the metatable into :New( Was the original plan for that. I had attached them to the class for the idea of using custom metatable for dropdowns but it never panned out.